Magellan: remove redundant casts
(Based on compiler warnings.)

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@15901 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ plugins/Magellan/src/org/micromanager/plugins/magellan/acq/FixedAreaAcquisition.java
@@ -140,8 +140,8 @@ public class FixedAreaAcquisition extends Acquisition {
          } else {
             //no space mode, use current stage positon
             positions_ = new ArrayList<XYStagePosition>();
-            int fullTileWidth = (int) JavaLayerImageConstructor.getInstance().getImageWidth();
-            int fullTileHeight = (int) JavaLayerImageConstructor.getInstance().getImageHeight();
+            int fullTileWidth = JavaLayerImageConstructor.getInstance().getImageWidth();
+            int fullTileHeight = JavaLayerImageConstructor.getInstance().getImageHeight();
             int tileWidthMinusOverlap = fullTileWidth - this.getOverlapX();
             int tileHeightMinusOverlap = fullTileHeight - this.getOverlapY();
             Point2D.Double currentStagePos = Magellan.getCore().getXYStagePosition(xyStage_);
